refactor(documentStepper): move stepper in a dedicated component
